credit-scoring-service/requirements.txt
fastapi
requests
uvicorn
pydantic
pandas
scikit-learn
joblib
numpy
pytests